requirements.txt
vosk
sounddevice
selenium
fuzzywuzzy
simpleaudio
pyaudio
pyyaml
pyinstaller
python-Levenshtein